graph ai system/backend/requirements.txt
fastapi
uvicorn
networkx
pandas
google-generativeai
